+++ resources/js/app.js
@@ -48,15 +48,15 @@ import Popper from "vue3-popper";
 
 // Vuetify
 // import '@mdi/font/css/materialdesignicons.css'
-import 'vuetify/styles'
-import { createVuetify } from 'vuetify'
-import * as components from 'vuetify/components'
-import * as directives from 'vuetify/directives'
-
-const vuetify = createVuetify({
-    components,
-    directives
-})
+// import 'vuetify/styles'
+// import { createVuetify } from 'vuetify'
+// import * as components from 'vuetify/components'
+// import * as directives from 'vuetify/directives'
+//
+// const vuetify = createVuetify({
+//     components,
+//     directives
+// })
 
 const appName = window.document.getElementsByTagName('title')[0]?.innerText || 'notTV';
 
@@ -94,7 +94,7 @@ createInertiaApp({
             .use(ZiggyVue, Ziggy)
             .use(createPinia())
             .use(Vue3TouchEvents)
-            .use(vuetify)
+            // .use(vuetify)
             // .use(vueCountryRegionSelect)
             // .use(setupCalendar, {})
             .component("Link", Link)